requirements.txt
setuptools==60.2.0
numpy==1.22.3
gpytorch==1.9.0
botorch==0.7
git+https://github.com/samuelstanton/upcycle.git@master
wandb==0.12.21
biopython==1.79
tabulate==0.8.9
pypdb==2.0
aiohttp==3.7.4
aioredis==1.3.0
omegaconf==2.1.0.dev26
hydra-core==1.1.0.dev6
cloudpickle==1.6.0
pickle5==0.0.11
tqdm==4.62.3
pandas==1.3.0
jupyter==1.0.0
seaborn==0.11.2